Amendment 39 2075-08-10
Objectives

Amendment designed to negate the judgement of Allahabad High Court invalidating Prime Minister Indira Gandhi's election to parliament. Amendment placed restrictions on judicial scrutiny of post of Prime Minister. The amendment was introduced and passed in the Lok Sabha on 7 August 1975 and again introduced and passed in the Rajya Sabha on 8 August 1975. As many as 17 State Assemblies, summoned on Saturday, 9 August ratified this amendment and President Fakhruddin Ali Ahmad gave his assent on Sunday, 10 August and the civil servants issued gazette notification on Sunday, 10 August 1975. As a consequence of this amendment to the Constitution of India, Supreme Court of India's scheduled hearing on 11 August 1975 of Petition challenging Prime Minister Indira Gandhi's election became infructuous.[44]Later however, Article 329A was struck down by the Supreme Court in case of State of Uttar Pradesh v. Raj Narain 1976 (2) SCR 347, for being in violation of basic structure.

Amendments

Amend articles 71 and 329.
Insert article 329A.
Amend schedule 9.[43]

Closes #43

Signed-off-by: sushmit1 <sushmit718@gmail.com>