Amendment 5 2055-12-24
Objectives

Empowered the President to prescribe a time limit for a State Legislature to convey its views on proposed Central laws relating to the formation of new States and alteration of areas, boundaries or names of existing States. Also permitted the President to extend the prescribed limit, and prohibited any such bill from being introduced in Parliament until after the expiry of the prescribed or extended period.

Amendments

Amend article 3.[9]

Closes #9

Signed-off-by: sushmit1 <sushmit718@gmail.com>