Amendment 52 2085-02-15
Objectives

Anti Defection Law - Provide disqualification of members from parliament and assembly in case of defection from one party to other. However, parts of the 10th Schedule to the Constitution of India was struck down by the Supreme Court in the case of Kihoto Hollohan v. Zachillhu 1992 SCR (1) 686, for being in contravention with Article 368 of the Constitution.

Amendments

Amend articles 101, 102, 190 and 191.
Insert schedule 10.[57]

Closes #56

Signed-off-by: sushmit1 <sushmit718@gmail.com>